grub-conf: Boot using filesystem UUIDS instead of labels
The filesystem UUIDs are inserted by the resin-image-flasher image in the
flashing process, or are read from a os_fs_uuids.txt file on the partition
grub boots from.

Note that this change make resin-image non-bootable if not installed into
the device via the resin-image-flasher image.

Changelog-entry: Boot using filesystem UUIDs instead of labels
Signed-off-by: Alex Gonzalez <alexg@balena.io>